refactor(30-A): delete V1 ITenantProvisioner surface (Wave C); 0 grep hits
Behavior port analysis: V1's CranlProvisioningWorkflow tests covered 5
behaviors (env-var payload, encrypted-conn round-trip, teardown delete
order, 404-as-absent, ShortenForName). None can be ported to
CranlTenantProviderV2Tests because CranlTenantProviderV2 has no
ICranlApiClient dependency — it is a thin dispatcher, not a Cranl API
walker. V2 behavior already covered by its existing test suite.

Files deleted (8 total):
- ITenantProvisioner.cs, NullTenantProvisioner.cs,
  CranlTenantProvisioner.cs, CranlProvisioningWorkflow.cs,
  TenantProvisioningTaskHandler.cs
- NullTenantProvisionerTests.cs, CranlTenantProvisionerTests.cs,
  CranlProvisioningWorkflowTests.cs

Collateral fixes:
- ProvisioningTaskPayload moved CranlTenantProvisioner → ProvisioningModels.cs
- ProvisioningTaskType/DeprovisioningTaskType constants moved to
  CranlTenantProviderV2 (same string values preserved)
- ProvisioningOptions + ProvisioningStatus records deleted from ProvisioningModels.cs
- All CS0618 pragma blocks removed (no more [Obsolete] references)
- Dangling crefs fixed in Tenant.cs, ProvisioningResult.cs,
  ITenantConnectionResolver.cs, CranlTenantProviderV2.cs,
  ProvisioningServiceCollectionExtensions.cs, CranlTenantProviderV2Tests.cs

Gates: build 0 errors / 508 warnings; tests 3771 passed, 0 failed, 6 skipped
