+++ apps/tamma-elsa/src/Tamma.Api/Extensions/ProvisioningServiceCollectionExtensions.cs
@@ -5,7 +5,6 @@ using Tamma.Api.Services.Provisioning;
 using Tamma.Api.Services.Provisioning.Cranl;
 using Tamma.Api.Services.Provisioning.V2;
 using Tamma.Api.Services.Provisioning.V2.Cranl;
-using Tamma.Api.Services.TaskQueue;
 
 namespace Tamma.Api.Extensions;
 
@@ -20,18 +19,16 @@ namespace Tamma.Api.Extensions;
 public static class ProvisioningServiceCollectionExtensions
 {
     /// <summary>
-    /// Register <see cref="ITenantProvisioner"/> + collaborators.
+    /// Register provisioning collaborators (Wave C — V1 surface removed).
     /// <list type="bullet">
     ///   <item><description>When <c>Cranl:ApiKey</c> + <c>Cranl:OrganizationId</c>
-    ///     are both set: registers <see cref="CranlTenantProvisioner"/>,
-    ///     <see cref="CranlProvisioningWorkflow"/>, the typed
-    ///     <see cref="ICranlApiClient"/> (with the auth header bound on
-    ///     the pooled HttpClient), the
-    ///     <see cref="TenantProvisioningTaskHandler"/>, and the
+    ///     are both set: registers the typed <see cref="ICranlApiClient"/>
+    ///     (with the auth header bound on the pooled HttpClient), the
     ///     <see cref="TenantSecretProtector"/> for at-rest encryption of
-    ///     the tenant DATABASE_URL.</description></item>
-    ///   <item><description>Otherwise: registers
-    ///     <see cref="NullTenantProvisioner"/> only — no external
+    ///     the tenant DATABASE_URL, and
+    ///     <see cref="CranlTenantProviderV2"/>.</description></item>
+    ///   <item><description>Otherwise: only the null-seam
+    ///     <see cref="NullTenantProvider"/> is registered — no external
     ///     resources are minted for any tenant and no Cranl wiring is
     ///     created.</description></item>
     /// </list>
@@ -52,9 +49,8 @@ public static class ProvisioningServiceCollectionExtensions
 
         if (options.IsConfigured)
         {
-            // Real Cranl path — needs the typed HttpClient, the workflow,
-            // the handler that the queue dispatches into, and the
-            // secret protector for DATABASE_URL encryption.
+            // Real Cranl path — needs the typed HttpClient, the v2 Cranl
+            // provider, and the secret protector for DATABASE_URL encryption.
             services.AddHttpClient<ICranlApiClient, CranlApiClient>((sp, client) =>
             {
                 var opts = sp.GetRequiredService<CranlOptions>();
@@ -70,42 +66,17 @@ public static class ProvisioningServiceCollectionExtensions
                 var env = sp.GetService<Microsoft.Extensions.Hosting.IHostEnvironment>();
                 return TenantSecretProtector.FromConfiguration(cfg, env, logger);
             });
-            services.TryAddScoped<CranlProvisioningWorkflow>();
-#pragma warning disable CS0618 // v1 surface is [Obsolete] (Story 30-3) but still wired
-                               // alongside v2 until wave-C migrates callers.
-            services.TryAddScoped<ITenantProvisioner, CranlTenantProvisioner>();
-#pragma warning restore CS0618
-            // Register handler under both ITaskHandler (so the registry sees
-            // it) and itself (so direct resolution works in tests).
-            services.AddScoped<ITaskHandler, TenantProvisioningTaskHandler>();
 
             // ── Story 30-3: v2 Cranl provider ──────────────────────────────
-            // Wired alongside v1 so both surfaces are available during the
-            // wave-B → wave-C transition. v2 takes the same dependencies
-            // (DbContext, platform queue repo, secret protector, options)
-            // and preserves the platform-queue dispatch pattern v1 relies on.
             services.AddTenantProviderCranl();
-        }
-        else
-        {
-#pragma warning disable CS0618 // v1 surface is [Obsolete] (Story 30-3) but still wired
-                               // alongside v2 until wave-C migrates callers.
-            services.TryAddScoped<ITenantProvisioner, NullTenantProvisioner>();
-#pragma warning restore CS0618
         }
 
         // ── Story 30-1: v2 ITenantInfrastructureProvider registry ───────────
         //
-        // The v2 surface coexists with the v1 ITenantProvisioner above. Story
-        // 30-3 will retire the v1 path and refactor CranlTenantProvisioner
-        // into a v2 provider; until then the v2 registry has only the null
-        // seam wired so the dispatch workflow + onboarding UI can take a
-        // hard dependency on TenantProviderRegistry without breaking
-        // existing single-user deployments. Mode behaviour:
+        // Mode behaviour:
         //   single-user: NullTenantProvider only — provisioning is unused.
-        //   SaaS:        NullTenantProvider + (eventually) per-backend
-        //                providers from 30-3..30-6, plugged in by their
-        //                own AddTenantProvider* extension methods.
+        //   SaaS:        NullTenantProvider + per-backend providers plugged
+        //                in by their own AddTenantProvider* extension methods.
         //
         // Note: providers are registered as IEnumerable<ITenantInfrastructureProvider>
         // — the TenantProviderRegistry consumes the collection at startup.
@@ -144,8 +115,7 @@ public static class ProvisioningServiceCollectionExtensions
     /// test fixture / future wave-C composition root can wire it
     /// independently of the v1 path.
     ///
-    /// <para><b>Lifetime</b>: scoped, mirroring v1's
-    /// <see cref="CranlTenantProvisioner"/>. The provider's only
+    /// <para><b>Lifetime</b>: scoped. The provider's only
     /// not-thread-safe dependency is the scoped
     /// <see cref="ControlPlaneDbContext"/>; everything else
     /// (<see cref="IPlatformQueuedTaskRepository"/>,
