--- apps/tamma-elsa/src/Tamma.Api/Services/Provisioning/ITenantProvisioner.cs
@@ -1,48 +1,0 @@
-namespace Tamma.Api.Services.Provisioning;
-
-/// <summary>
-/// Per-tenant provisioning surface. Two implementations:
-/// <list type="bullet">
-///   <item><description><see cref="NullTenantProvisioner"/> — dev/default
-///     when <c>Cranl:ApiKey</c> is absent. Flips the row to
-///     <see cref="ProvisioningState.Ready"/> immediately; the tenant's
-///     placement stays on the unified tenant_databases pool. No external
-///     calls.</description></item>
-///   <item><description><see cref="CranlTenantProvisioner"/> — production.
-///     Walks the README's per-tenant flow (project → db → poll → app
-///     → env → deploy → poll → domains).</description></item>
-/// </list>
-///
-/// <para>All methods are idempotent: calling <see cref="ProvisionAsync"/>
-/// on a tenant that already has a Cranl project returns the current
-/// status without doing anything new.</para>
-///
-/// <para><b>Story 30-3 — DEPRECATED.</b> Replaced by the v2 surface
-/// <see cref="V2.ITenantInfrastructureProvider"/> + the
-/// <see cref="V2.TenantProviderRegistry"/>. Existing call sites
-/// (admin endpoints, the platform task handler) continue to compile —
-/// the wave-C migration converts them to the v2 contract. New code
-/// MUST use v2; no exceptions.</para>
-/// </summary>
-[Obsolete("Use ITenantInfrastructureProvider (V2) instead. Removed in Wave C.")]
-public interface ITenantProvisioner
-{
-    /// <summary>
-    /// Trigger provisioning for the tenant. Returns immediately with
-    /// <see cref="ProvisioningState.Pending"/> (or the current state if
-    /// the tenant already has a project) — the long-running Cranl
-    /// polling work runs on a background task.
-    /// </summary>
-    Task<ProvisioningStatus> ProvisionAsync(
-        Guid tenantId, ProvisioningOptions options, CancellationToken ct = default);
-
-    /// <summary>Read the current provisioning state for the tenant.</summary>
-    Task<ProvisioningStatus> GetStatusAsync(Guid tenantId, CancellationToken ct = default);
-
-    /// <summary>
-    /// Tear down the tenant's Cranl resources. Sequence: delete app, then
-    /// db, then project (per Cranl's "project must have no apps before
-    /// delete" constraint). Clears all <c>cranl_*</c> columns on success.
-    /// </summary>
-    Task DeprovisionAsync(Guid tenantId, CancellationToken ct = default);
-}
