+++ apps/tamma-elsa/src/Tamma.Api/Services/Provisioning/ProvisioningModels.cs
@@ -85,12 +85,14 @@ public static class ProvisioningStateExtensions
     };
 }
 
-/// <summary>Caller-supplied options for kicking off provisioning.</summary>
-public sealed record ProvisioningOptions(string Region, string? CustomName = null);
-
-/// <summary>Snapshot of the current provisioning state for a tenant.</summary>
-public sealed record ProvisioningStatus(
-    ProvisioningState State,
-    string? Detail,
-    string? AppDefaultDomain,
-    DateTimeOffset UpdatedAt);
+/// <summary>
+/// Payload serialised into a platform-queue task row for provisioning and
+/// deprovisioning operations. Consumed by <see cref="V2.Cranl.CranlTenantProviderV2"/>
+/// when enqueueing background work on the platform queue.
+/// </summary>
+public sealed class ProvisioningTaskPayload
+{
+    public Guid TenantId { get; set; }
+    public string Region { get; set; } = string.Empty;
+    public string? CustomName { get; set; }
+}
